reduce node size

+++ components/script/dom/html/htmlcollection.rs
@@ -84,7 +84,7 @@ pub(crate) struct HTMLCollection {
     // We cache the version of the root node and all its decendents,
     // the length of the collection, and a cursor into the collection.
     // FIXME: make the cached cursor element a weak pointer
-    cached_version: Cell<u64>,
+    cached_version: Cell<u32>,
     cached_cursor_element: MutNullableDom<Element>,
     cached_cursor_index: Cell<OptionU32>,
     cached_length: Cell<OptionU32>,
